SEARCH-1003 Fix a couple of issues raised by Sonar.
Replace some code that throws a NullPointerException with code to explicitly throw
an AlfrescoRuntimeException if the core name is not passed when creating the
transaction report.

Simplify the logic when getting the first live doc for a search term by not reusing
the postings iterator (nb. It wasn't being reused anyway as we always passed null).